+++ CHANGELOG.md
@@ -6,6 +6,7 @@
     -   [is_bool](./docs/variables/is_bool.md)
     -   [is_callable](./docs/variables/is_callable.md)
     -   [is_class](./docs/variables/is_class.md)
+    -   [is_contract](./docs/variables/is_contract.md)
     -   [is_object](./docs/variables/is_object.md)
 
 -   Добавлена функция [array](./docs_depricated/funcref/vartype/array/func/array.md) в модуль